requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Werkzeug==3.1.3
python-dotenv==1.0.1
email-validator==2.2.0
